+++ Chilano/Iso2God/ConStructures/SubHashTable.cs
@@ -19,7 +19,7 @@
             }
             foreach (byte[] buffer2 in this)
             {
-                buffer2.CopyTo(array, (long) num);
+                buffer2.CopyTo(array, num);
                 num += 20;
             }
             return array;
